change handling of tumbling window without end as it won't work with variable stack offsets
We could not call the start twice, since the first evaluation adds variables, changing the offsets. We could not drop the variables from the first evaluation as they are needed for the later subclauses. We could not copy the start construction to an end construction, because then it would have duplicate variables. We cannot ignore those variables as they change the offset of the later subclauses. It might have been possible to rename them though