fix last commit
variety is never really used, and left at the default setting. if the default is absent, the schema initialization fails because you cannot create a list (NMTOKENS) of a type with absent variety (NMTOKEN which should have atomic variety). Before the previous commit the list check was to lenient and allowed lists of types with absent variety